Remove deprecated BEL/RDF-based tests
